editevent refresh process is going
